Refactor task runtime lineage model
Remove legacy columns (`child_thread_id`, `child_turn_id`,
`parent_turn_id`,
`task_id`, `task_run_id`) from `thread_lineage` and
`task_run_execution`.
Replace with `task_run_thread_binding` and `task_run_turn`
relationships.
Introduce `TaskThreadLineage` protocol type and
`TaskThreadLineageCreated`
event. Add migration to drop old indexes and columns after data
validation.
Update invariant checks to use the new snapshot-based approach.

+++ crates/entity/src/task_result_candidate.rs
@@ -10,12 +10,14 @@ pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
     pub id: String,
     pub task_id: String,
+    #[sea_orm(unique)]
     pub run_id: String,
     #[sea_orm(unique)]
     pub task_run_turn_id: String,
     pub thread_id: String,
     pub turn_id: String,
-    pub round: i32,
+    #[sea_orm(unique_key = "uidx_task_result_candidate_run_round")]
+    pub round: i64,
     pub status: String,
     #[sea_orm(column_type = "Text", nullable)]
     pub result_json: Option<String>,
